ECDC International Case Counts

ECDC International Case Counts

get_ecdc_cases(countries = NULL)

Arguments

countries
Character vector identifying the countries to extract data for.

Value

A dataframe of International case counts published by ECDC.

Examples

## Get data for france
get_ecdc_cases(countries = "France")

#> # A tibble: 79 x 5
#>    date       cases deaths country geoid
#>    <date>     <dbl>  <int> <fct>   <fct>
#>  1 2019-12-31     0      0 France  FR   
#>  2 2020-01-01     0      0 France  FR   
#>  3 2020-01-02     0      0 France  FR   
#>  4 2020-01-03     0      0 France  FR   
#>  5 2020-01-04     0      0 France  FR   
#>  6 2020-01-05     0      0 France  FR   
#>  7 2020-01-06     0      0 France  FR   
#>  8 2020-01-07     0      0 France  FR   
#>  9 2020-01-08     0      0 France  FR   
#> 10 2020-01-09     0      0 France  FR   
#> # … with 69 more rows

## Code
get_ecdc_cases

#> function (countries = NULL) 
#> {
#>     base_url = "http://www.ecdc.europa.eu/sites/default/files/documents/COVID-19-geographic-disbtribution-worldwide-"
#>     ch <- memoise::cache_filesystem(".cache")
#>     mem_read <- memoise::memoise(gdata::read.xls, cache = ch)
#>     try_dates <- c(as.Date(Sys.time()) - 1, as.Date(Sys.time()))
#>     for (i in 1:length(try_dates)) {
#>         date <- try_dates[i]
#>         url <- paste0(base_url, date, ".xls")
#>         d <- try(tibble::as_tibble(mem_read(url)), silent = TRUE)
#>         if (length(d) > 1) {
#>             break
#>         }
#>     }
#>     if (length(d) == 1) {
#>         stop(paste0("No data found at: ", url, "\n Tried dates: ", 
#>             try_dates[1], ", ", try_dates[2]))
#>     }
#>     d <- d %>% dplyr::mutate(date = as.Date(DateRep), cases = Cases, 
#>         deaths = Deaths, country = Countries.and.territories, 
#>         geoid = GeoId) %>% dplyr::select(date, cases, deaths, 
#>         country, geoid) %>% dplyr::arrange(date) %>% dplyr::mutate(cases = ifelse(cases < 
#>         0, 0, cases))
#>     if (!is.null(countries)) {
#>         d <- d %>% dplyr::filter(country %in% countries)
#>     }
#>     return(d)
#> }
#> <bytecode: 0x5563fefa4cd0>
#> <environment: namespace:NCoVUtils>