Get Total Cases by Country

Get Total Cases by Country

get_total_cases(source = "WHO")

Arguments

source
Character String specifying data source for total cases "WHO", "ECDC", default: "WHO".

Value

Returns a data frame of variables along with their most recent value.

Examples

## Total cases sourced from the WHO
get_total_cases(source = "WHO")

#> Warning: 'autostart' is now deprecated and ignored. Consider skip='string' or skip=n

#> Warning: NAs introduced by coercion

#> # A tibble: 230 x 2
#>    country               cases
#>    <chr>                 <dbl>
#>  1 China                 81416
#>  2 Italy                 47021
#>  3 Spain                 19980
#>  4 Iran                  19644
#>  5 Germany               18323
#>  6 UnitedStatesofAmerica 15219
#>  7 France                12475
#>  8 RepublicofKorea        8799
#>  9 Switzerland            4840
#> 10 UnitedKingdom          3983
#> # … with 220 more rows

## Total cases sourced from ECDC
get_total_cases(source = "ECDC")

#> # A tibble: 180 x 2
#>    country                  cases
#>    <chr>                    <dbl>
#>  1 China                    81499
#>  2 Italy                    53578
#>  3 United_States_of_America 26747
#>  4 Spain                    24926
#>  5 Germany                  21463
#>  6 Iran                     20610
#>  7 France                   14459
#>  8 South_Korea               8897
#>  9 Switzerland               6077
#> 10 United_Kingdom            5018
#> # … with 170 more rows