ECDC International Case Counts

ECDC International Case Counts

get_ecdc_cases(countries = NULL)

Arguments

countries
Character vector identifying the countries to extract data for.

Value

A dataframe of International case counts published by ECDC.

Examples

## Get data for france
get_ecdc_cases(countries = "France")

#> Warning: the condition has length > 1 and only the first element will be used

#> # A tibble: 90 x 10
#>      day month  year cases deaths country geoid countryterritor… population_2018
#>    <dbl> <dbl> <dbl> <dbl>  <dbl> <chr>   <chr> <chr>                      <dbl>
#>  1    31    12  2019     0      0 France  FR    FRA                     66987244
#>  2     1     1  2020     0      0 France  FR    FRA                     66987244
#>  3     2     1  2020     0      0 France  FR    FRA                     66987244
#>  4     3     1  2020     0      0 France  FR    FRA                     66987244
#>  5     4     1  2020     0      0 France  FR    FRA                     66987244
#>  6     5     1  2020     0      0 France  FR    FRA                     66987244
#>  7     6     1  2020     0      0 France  FR    FRA                     66987244
#>  8     7     1  2020     0      0 France  FR    FRA                     66987244
#>  9     8     1  2020     0      0 France  FR    FRA                     66987244
#> 10     9     1  2020     0      0 France  FR    FRA                     66987244
#> # … with 80 more rows, and 1 more variable: date <date>

## Code