Get Total Cases by Country

Get Total Cases by Country

get_total_cases(source = "WHO")

Arguments

source
Character String specifying data source for total cases "WHO", "ECDC", default: "WHO".

Value

Returns a data frame of variables along with their most recent value.

Examples

## Total cases sourced from the WHO
get_total_cases(source = "WHO")

#> Warning: 'autostart' is now deprecated and ignored. Consider skip='string' or skip=n

#> Warning: NAs introduced by coercion

#> # A tibble: 250 x 2
#>    country                cases
#>    <chr>                  <dbl>
#>  1 UnitedStatesofAmerica 103321
#>  2 Italy                  92472
#>  3 China                  82356
#>  4 Spain                  72248
#>  5 Germany                52547
#>  6 France                 37145
#>  7 Iran                   35408
#>  8 UnitedKingdom          17093
#>  9 Switzerland            13152
#> 10 Netherlands             9762
#> # … with 240 more rows

## Total cases sourced from ECDC
get_total_cases(source = "ECDC")

#> Warning: the condition has length > 1 and only the first element will be used

#> # A tibble: 196 x 2
#>    country                   cases
#>    <chr>                     <dbl>
#>  1 United_States_of_America 124665
#>  2 Italy                     92472
#>  3 China                     82342
#>  4 Spain                     72248
#>  5 Germany                   52547
#>  6 France                    37575
#>  7 Iran                      35408
#>  8 United_Kingdom            17089
#>  9 Switzerland               13152
#> 10 Netherlands                9762
#> # … with 186 more rows